fine_tune_package/requirements.txt
Flask
py2app
mlx
mlx_lm
transformers
numpy
pydantic
flask-cors
celery
redis
psycopg2-binary